Don't reuse the "ssh" group for our own access control
The "ssh" group was supposed to be an internal group that was used by
openssh itself for privilege separation. It's since been renamed to
"_ssh"[1] to clarify that it's internal-only and flagged that we
shouldn't be using it either.

Instead, create a new "sdssh" group, add our users to it, and use it in
the sshd_config.

For upgrades, we migrate users to the new group and then update
sshd_config and iptables rules.

[1] https://salsa.debian.org/ssh-team/openssh/-/blob/eac38305fc5d8eb8301a106294cf6c79447bdeb3/debian/openssh-client.postinst

Fixes #7316.