Remove haveged if installed
Previously we only disabled/masked haveged instead of removing the
package. Now that we have the infrastructure to remove packages (created
for ufw), let's remove it properly.

Note that this will remove it from both app and mon, even though it was
only ever disabled on app. That should be fine since it's not expected
to be running on mon.

Fixes #7326.