Switch systemd units to use Type=exec instead of Type=oneshot
`oneshot` is designed to be used with RemainAfterExit, but we don't need
that nor use it. After reviewing
<https://www.freedesktop.org/software/systemd/man/latest/systemd.service.html>,
`exec` appears to be the correct type since we want the service to be
running once the binary has started, not once it's finished.

Practically this shouldn't make a big difference except that `systemctl
start <name>` won't wait for the entire command to complete.

Fixes #7349.