Only try to apply grsec_lock once
Currently we specify both `sysctl_set: yes` and `reload: yes` when
setting sysctl settings, which ends up with it being applied twice,
first with `sysctl -w` (sysctl_set) and then through `sysctl -p`
(reload).

With noble/Linux 6.6, setting the lock twice errors out, so just enable
it once with `sysctl -p`. This is also closer to what the kernel will do
normally when booting in which the whole file is loaded at once.

Refs #7323.