Don't use the word "failed" in a success message; triggers OSSEC
The line "systemd OK: no failed units" was triggering an OSSEC
notification because "failed" is a bad word[1] that OSSEC alerts on.

So let's use the opportunity to focus on the positive instead of the
negative and explain that in the success case, all the units are happy.

A regression test verifies that piping the output into ossec-logtest
does not trigger any notifications. We already know that all the checks
should be passing in a staging environment because of the test in
test_release_upgrades.py. (The exemption for free space was a leftover
from before 87c9af4ce23, which fixed the calcuation.)

[1]
https://github.com/ossec/ossec-hids/blob/3.6.0/etc/rules/syslog_rules.xml#L21

Fixes #7393.