Don't use finally to return a redirect
As described in
<https://docs.astral.sh/ruff/rules/jump-statement-in-finally/> and more
recently in <https://peps.python.org/pep-0765/>, using return in a
finally block has the potential of suppressing exceptions.

Given that in both cases we have a blanket exception handler, it should
be safe to return outside of a finally. If there are exceptions being
suppressed, we should surface them.