Ensure /etc/iptables exists before writing to it
In the specific case of installing a fresh mon server when the app
server is already configured AND you're using ssh over the local
network, we'll try to write to /etc/iptables before the
iptables-persistent package is installed.

This is because we end up running the restrict-direct-access role before
the common role, which installs the base packages.

The easy fix is to install iptables-persistent ahead of time if we see
that it's necessary.

Fixes #7119.